+++ weaver/wps_restapi/providers/providers.py
@@ -2,6 +2,7 @@ import logging
 from typing import TYPE_CHECKING
 
 import colander
+from box import Box
 from pyramid.httpexceptions import (
     HTTPBadRequest,
     HTTPConflict,
@@ -15,7 +16,7 @@ from pyramid.settings import asbool
 from weaver.database import get_db
 from weaver.datatype import Process, Service
 from weaver.exceptions import ServiceNotFound, ServiceParsingError, log_unhandled_exceptions
-from weaver.formats import ContentType, OutputFormat
+from weaver.formats import ContentType, OutputFormat, guess_target_format
 from weaver.owsexceptions import OWSMissingParameterValue, OWSNotImplemented
 from weaver.processes.execution import submit_job
 from weaver.store.base import StoreServices
@@ -175,7 +176,16 @@ def get_provider(request):
     return HTTPOk(json=data)
 
 
-# FIXME: Add HTML view??? (same as local process, but extra 'provider' field?)
+@sd.provider_processes_service.get(
+    tags=[sd.TAG_PROVIDERS, sd.TAG_PROCESSES, sd.TAG_GETCAPABILITIES],
+    schema=sd.ProviderProcessesEndpoint(),
+    accept=ContentType.TEXT_HTML,
+    renderer="weaver.wps_restapi:templates/responses/process_listing.mako",
+    response_schemas=sd.derive_responses(
+        sd.get_provider_processes_responses,
+        sd.GenericHTMLResponse(name="HTMLProcessListing", description="Listing of processes.")
+    ),
+)
 @sd.provider_processes_service.get(
     tags=[sd.TAG_PROVIDERS, sd.TAG_PROCESSES, sd.TAG_GETCAPABILITIES],
     schema=sd.ProviderProcessesEndpoint(),
@@ -200,6 +210,16 @@ def get_provider_processes(request):
     links = get_process_list_links(request, paging={}, total=None, provider=service)
     body = {"processes": processes, "links": links}
     body = sd.ProcessesListing().deserialize(body)
+    ctype = guess_target_format(request)
+    if ctype == ContentType.TEXT_HTML:
+        # remap to properties expected by HTML template
+        # it assumes that both local/remote processes can be returned
+        # therefore, remove the "local" processes to list only the current provider
+        provider_processes = body.pop("processes")
+        body["providers"] = [{"id": service.id, "url": service.url, "processes": provider_processes}]
+        body["processes"] = []
+        body["total"] = len(provider_processes)
+        return Box(body)
     return HTTPOk(json=body)
 
 
@@ -221,6 +241,16 @@ def describe_provider_process(request, provider_id=None):
 
 @sd.provider_process_service.get(
     tags=[sd.TAG_PROVIDERS, sd.TAG_PROCESSES, sd.TAG_DESCRIBEPROCESS],
+    schema=sd.ProviderProcessesEndpoint(),
+    accept=ContentType.TEXT_HTML,
+    renderer="weaver.wps_restapi:templates/responses/process_description.mako",
+    response_schemas=sd.derive_responses(
+        sd.get_provider_process_responses,
+        sd.GenericHTMLResponse(name="HTMLProcessDescription", description="Process description.")
+    ),
+)
+@sd.provider_process_service.get(
+    tags=[sd.TAG_PROVIDERS, sd.TAG_PROCESSES, sd.TAG_DESCRIBEPROCESS],
     accept=ContentType.APP_JSON,
     renderer=OutputFormat.JSON,
     schema=sd.ProviderProcessEndpoint(),
@@ -234,10 +264,11 @@ def get_provider_process(request):
     """
     Retrieve a remote provider's process description (DescribeProcess).
     """
-    process, _ = describe_provider_process(request)
+    process, service = describe_provider_process(request)
     schema = request.params.get("schema")
     offering = process.offering(schema)
-    return HTTPOk(json=offering)
+    offering["provider"] = service.summary(request)
+    return Box(offering)
 
 
 @sd.provider_process_package_service.get(